# pharma-platform/server/docs/openapi.yaml
openapi: 3.1.0
info:
  title: PharmaGuard API
  version: 1.0.0
  description: |
    Pharmaceutical Data Integrity & Compliance Platform.
    FDA 21 CFR Part 11 compliant electronic records, audit trails, 
    e-signatures, batch tracking, and document control APIs.
  contact:
    name: PharmaGuard Engineering
    email: api@pharmaguard.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api/v1
    description: Local Development
  - url: https://api.pharmaguard.io/api/v1
    description: Production
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for watch agents and external integrations
  schemas:
...
    FileUpload:
      type: object
      properties:
        id: { type: string, format: uuid }
        originalFilename: { type: string, example: "batch_001_hplc_run.cdf" }
        fileSize: { type: integer, example: 2048576 }
        sha256Hash: { type: string, example: "a1b2c3d4e5f6..." }
        md5Hash: { type: string }
        status: 
          type: string
          enum: [pending_verification, verified, flagged, quarantined, archived]
        integrityVerified: { type: boolean }
        instrumentId: { type: string, format: uuid }
        batchId: { type: string, format: uuid }
        uploadedBy: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
paths:
...
  /uploads:
    get:
      tags: [Uploads]
      summary: List uploaded files
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: string } }
        - { name: instrumentId, in: query, schema: { type: string, format: uuid } }
        - { name: batchId, in: query, schema: { type: string, format: uuid } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200':
          description: Paginated list of uploads
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { type: array, items: { $ref: '#/components/schemas/FileUpload' } }
                  pagination: { $ref: '#/components/schemas/Pagination' }

    post:
      tags: [Uploads]
      summary: Upload instrument data file
      description: |
        Uploads a file with automatic SHA-256 and MD5 hashing for data integrity.
        Duplicate files (same hash) are rejected.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                instrumentId: { type: string, format: uuid }
                batchId: { type: string, format: uuid }
                notes: { type: string }
      responses:
        '201': { description: File uploaded and verified }
        '409': { description: Duplicate file detected }
  /uploads/{id}:
    get:
      tags: [Uploads]
      summary: Get upload details
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Upload details }
  /uploads/{id}/flag:
    patch:
      tags: [Uploads]
      summary: Flag file for review
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '200': { description: File flagged }
  /uploads/{id}/verify:
    patch:
      tags: [Uploads]
      summary: Re-verify file integrity
      description: Re-computes SHA-256 hash and compares against stored hash
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  isValid: { type: boolean }
                  expectedHash: { type: string }
                  actualHash: { type: string }

  # ─── Audit Trail ───────────────────────────────────────
...